Move mlt_image_box_blur to module private file image_proc
