+++ src/modules/core/factory.c
@@ -43,6 +43,7 @@ extern mlt_filter filter_mirror_init( mlt_profile profile, mlt_service_type type
 extern mlt_filter filter_mono_init( mlt_profile profile, mlt_service_type type, const char *id, char *arg );
 extern mlt_filter filter_obscure_init( mlt_profile profile, mlt_service_type type, const char *id, char *arg );
 extern mlt_filter filter_panner_init( mlt_profile profile, mlt_service_type type, const char *id, char *arg );
+extern mlt_filter filter_pillar_echo_init( mlt_profile profile, mlt_service_type type, const char *id, char *arg );
 extern mlt_filter filter_rescale_init( mlt_profile profile, mlt_service_type type, const char *id, char *arg );
 extern mlt_filter filter_resize_init( mlt_profile profile, mlt_service_type type, const char *id, char *arg );
 extern mlt_filter filter_transition_init( mlt_profile profile, mlt_service_type type, const char *id, char *arg );
@@ -95,6 +96,7 @@ MLT_REPOSITORY
 	MLT_REGISTER( mlt_service_filter_type, "mono", filter_mono_init );
 	MLT_REGISTER( mlt_service_filter_type, "obscure", filter_obscure_init );
 	MLT_REGISTER( mlt_service_filter_type, "panner", filter_panner_init );
+	MLT_REGISTER( mlt_service_filter_type, "pillar_echo", filter_pillar_echo_init );
 	MLT_REGISTER( mlt_service_filter_type, "rescale", filter_rescale_init );
 	MLT_REGISTER( mlt_service_filter_type, "resize", filter_resize_init );
 	MLT_REGISTER( mlt_service_filter_type, "transition", filter_transition_init );
@@ -140,6 +142,7 @@ MLT_REPOSITORY
 	MLT_REGISTER_METADATA( mlt_service_filter_type, "mono", metadata, "filter_mono.yml" );
 	MLT_REGISTER_METADATA( mlt_service_filter_type, "obscure", metadata, "filter_obscure.yml" );
 	MLT_REGISTER_METADATA( mlt_service_filter_type, "panner", metadata, "filter_panner.yml" );
+	MLT_REGISTER_METADATA( mlt_service_filter_type, "pillar_echo", metadata, "filter_pillar_echo.yml" );
 	MLT_REGISTER_METADATA( mlt_service_filter_type, "rescale", metadata, "filter_rescale.yml" );
 	MLT_REGISTER_METADATA( mlt_service_filter_type, "resize", metadata, "filter_resize.yml" );
 	MLT_REGISTER_METADATA( mlt_service_filter_type, "transition", metadata, "filter_transition.yml" );
