+++ src/modules/core/filter_pillar_echo.c
@@ -17,6 +17,8 @@
  * Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301  USA
  */
 
+#include "image_proc.h"
+
 #include <framework/mlt.h>
 #include <framework/mlt_log.h>
