+++ src/modules/core/filter_pillar_echo.yml
