+++ src/modules/core/image_proc.h
@@ -1,0 +1,27 @@
+/*
+ * Copyright (c) 2022 Meltytech, LLC
+ *
+ * This library is free software; you can redistribute it and/or
+ * modify it under the terms of the GNU Lesser General Public
+ * License as published by the Free Software Foundation; either
+ * version 2.1 of the License, or (at your option) any later version.
+ *
+ * This library is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU
+ * Lesser General Public License for more details.
+ *
+ * You should have received a copy of the GNU Lesser General Public
+ * License along with this library; if not, write to the Free Software
+ * Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301  USA
+ */
+
+#ifndef IMAGE_PROC_H
+#define IMAGE_PROC_H
+
+#include <framework/mlt_events.h>
+#include <framework/mlt_image.h>
+
+void mlt_image_box_blur( mlt_image self, int hradius, int vradius );
+
+#endif // IMAGE_PROC_H
